Add language models and DbContext setup
This commit introduces five new classes under the namespace Getig.Models.Language: Language, Syllable, OrthographyRule, Word, and Phoneme. These models represent various components of a spoken language for a language processing tool. Each model class has properties to define it and has connections with other models.

In the Getig.Services namespace, we have a new dbContext class set up - GetigDbContext. This class represents a session with the database enabling it to perform operations like CRUD. It included the configuration for Entity Frameworks core to interact with a Microsoft Access database.

Also, a new service class ProcLangService which is planned to handle language processing is added.

Changes are done to the Getig.Services/Project.cs file to make use of the dbContext.

In Getig.Service.csproj, new package references for EntityFrameworkCore.Jet and Microsoft.EntityFrameworkCore have been added. The changes add capabilities to interact with Jet databases with the Microsoft EntityFramework library to perform CRUD operations.