Implement Language model and introduce LanguageService and IRepository
This commit implements Language, a model representing a spoken language with collections of Phonetic, Syllable, Word, and Orthography rules. Also, includes a language service: LanguageService and a generic repository interface IRepository to provide manipulation methods like Add, GetById, and Update for the domain model.

These changes are aimed to set up the foundational model and the data access layer for the project, facilitating fluent communication between different layers of the application. Updates on the UI were also performed to reflect these changes and provide a better way of communication for the end-users with the application.

The model classes Language, Phonetic, Syllable, Word, and Orthography Rule were modified to align with the new design. The Program.cs and CsProj were also adjusted to accommodate these modifications. The database context class, GetigDbContext, is updated to include references for all the new models, and migrations need to be run after this update to reflect these changes in the database schema.