Add new language service and update DB context path
The modifications include the addition of new LanguageService with the method to add a constructed language. The GetigDbContext's relative path has been updated for its location to be dynamically determined at runtime. New lookup classes have also been added. The changes aim to enhance the language construction functionality and maintain the flexibility of the database file location.